+++ package.json
@@ -17,6 +17,7 @@
         "@types/cookie": "^0.5.1",
         "@typescript-eslint/eslint-plugin": "^5.45.0",
         "@typescript-eslint/parser": "^5.45.0",
+        "drizzle-kit": "^0.22.7",
         "eslint": "^8.28.0",
         "eslint-config-prettier": "^8.5.0",
         "eslint-plugin-svelte": "^2.30.0",
@@ -39,6 +40,8 @@
         "date-picker-svelte": "^2.6.0",
         "dayjs": "^1.11.9",
         "dotenv": "^16.4.5",
+        "drizzle-orm": "^0.31.2",
+        "postgres": "^3.4.4",
         "radix-icons-svelte": "^1.2.1"
     }
 }
