+++ src/lib/supabase.ts
@@ -8,17 +8,3 @@ export const supabase = createClient(
     "https://bbwzezdaeqapoeyiuoyo.supabase.co",
     SUPABASE_ANON_KEY
 );
-
-export async function signInWithEmailMagicLink(email: string) {
-    const { error } = await supabase.auth.signInWithOtp({
-        email: email,
-        options: {
-            emailRedirectTo: window.location.origin + "?magiclink=true",
-        },
-    });
-    if (error) throw error;
-}
-
-export async function signOut() {
-    await supabase.auth.signOut();
-}
